requirements.txt
chispa==0.11.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
delta-spark==3.3.2
exceptiongroup==1.3.1
fonttools==4.61.1
importlib_metadata==8.7.1
iniconfig==2.3.0
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pluggy==1.6.0
prettytable==3.17.0
py4j==0.10.9.7
Pygments==2.19.2
pyparsing==3.3.1
pyspark==3.5.3
pytest==9.0.2
python-dateutil==2.9.0.post0
pytz==2025.2
seaborn==0.13.2
six==1.17.0
tomli==2.4.0
typing_extensions==4.15.0
tzdata==2025.3
wcwidth==0.2.14
zipp==3.23.0
